backend/requirements.txt
fastapi
uvicorn[standard]
websockets
google-generativeai
pydantic-settings